pyproject.toml
[project]
name = "vibecodeinfo"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "asyncio (>=4.0.0,<5.0.0)",
    "aiogram (>=3.22.0,<4.0.0)",
    "apscheduler (>=3.11.1,<4.0.0)",
    "newsapi-python (>=0.2.7,<0.3.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "aiojobs (>=1.4.0,<2.0.0)",
    "prometheus-client (>=0.23.1,<0.24.0)"
]

[dependency-groups]
dev = [
    "pre-commit (>=4.3.0,<5.0.0)"
]
lint = [
    "mypy (>=1.18.2,<2.0.0)",
    "flake8 (>=6.1.0,<7.0.0)",
]
tests = [
    "pytest (>=8.4.2,<9.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
